web/requirements.txt
flask
flask_restful
flask_sqlalchemy
elasticsearch~=7.17.4
python-dotenv~=0.20.0
inflection
requests
SQLAlchemy==1.4.35
psycopg2-binary==2.9.3
python-dotenv~=0.20.0
logzero==1.7.0
uvicorn==0.18.2
fastapi==0.78.0
strawberry-graphql==0.117.1
pydantic==1.9.1
asyncpg==0.26.0


